\name{gxeRC-package}
\alias{gxeRC-package}
\docType{package}
\title{
\packageTitle{gxeRC}
}
\description{
\packageDescription{gxeRC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gxeRC}
\packageIndices{gxeRC}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{gxeRC}

Maintainer: \packageMaintainer{gxeRC}
}
\references{
%~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
